Refactoring transformation and resampling code. Now they are on a separate class.
Refactoring and restructuring the factory of feature, replacing single function
with a proper Factory class.
Replace a simple list of feature with an enhanced class with more methods.

Three test has been marked as skipped to replace a previous TODO hidden in the code
and to highlight the fact that the test for classification with bin is missing.