Handle exceptions in processes pool context manager
We handle the mp.pool via a context manager to ensure that processes get
terminated (thus releasing memory) after they've completed their tasks.
If an exception gets raised from the function mapped on the iterable,
that exception will bubble up to the runner layer, but the context
manager won't make the processes release them. Having a try/finally
clause _inside_ the context manager makes sure that whatever happens in
the context manager we get to the line where the release should occur.